move: 패키지 구조 개선
- 각각의 클래스가 의미하는 패키지로 이동
