+++ src/main/java/be/codesquad/issuetracker/oauth/service/LoginService.java
@@ -1,9 +1,8 @@
-package be.codesquad.issuetracker.oauth;
+package be.codesquad.issuetracker.oauth.service;
 
 import be.codesquad.issuetracker.oauth.dto.GithubUser;
-import be.codesquad.issuetracker.user.User;
-import be.codesquad.issuetracker.user.UserRepository;
-import java.util.NoSuchElementException;
+import be.codesquad.issuetracker.user.domain.User;
+import be.codesquad.issuetracker.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
